refactor: move definitions
